requirements.txt
absl-py==2.3.1
actionlib==1.14.1
aiofiles==25.1.0
angles==1.9.14
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
bondpy==1.8.6
brax==0.14.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.11
certifi==2025.10.5
charset-normalizer==3.4.4
chex==0.1.90
click==8.3.0
colorama==0.4.6
commentjson==0.9.0
contourpy==1.3.2
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.4
etils==1.13.0
evdev==1.9.2
exceptiongroup==1.3.0
fastapi==0.121.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.2
flax==0.10.7
fonttools==4.60.1
fsspec==2025.9.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.1
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.17
glfw==2.10.0
h11==0.16.0
httptools==0.7.1
humanize==4.15.0
idna==3.11
image-geometry==1.16.2
importlib_resources==6.5.2
interactive-markers==1.12.1
itsdangerous==2.2.0
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jaxopt==0.8.5
Jinja2==3.1.6
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
kiwisolver==1.4.9
lark-parser==0.7.8
laser-geometry==1.6.8
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
message-filters==1.17.0
ml_collections==1.1.0
ml_dtypes==0.5.4
moveit-core==1.1.16
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.4.0
mujoco-mjx==3.4.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
numpy-stl==3.2.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
onshape_to_robot==1.7.9
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.31
packaging==25.0
pandas==2.3.3
pillow==11.3.0
playground==0.0.5
protobuf==6.33.2
psutil==7.2.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-qt-binding==0.4.5
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
qt-dotgraph==0.4.3
qt-gui==0.4.3
qt-gui-cpp==0.4.3
qt-gui-py-common==0.4.3
requests==2.32.5
resource-retriever==1.12.9
rich==14.2.0
robola==0.2.4
rosbag==1.17.0
rosboost-cfg==1.15.9
rosclean==1.15.9
roscreate==1.15.9
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.9
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.9
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.9
roswtf==1.17.0
rqt-action==0.4.10
rqt-bag==0.5.2
rqt-bag-plugins==0.5.2
rqt-console==0.4.13
rqt-dep==0.4.13
rqt-graph==0.4.15
rqt-gui==0.5.4
rqt-gui-py==0.5.4
rqt-image-view==0.4.18
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.11
rqt-plot==0.4.15
rqt-pose-view==0.5.12
rqt-publisher==0.4.11
rqt-py-common==0.5.4
rqt-py-console==0.4.11
rqt-reconfigure==0.5.6
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.13
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.1
rqt-service-caller==0.4.11
rqt-srv==0.4.10
rqt-tf-tree==0.6.4
rqt-top==0.4.11
rqt-topic==0.4.14
rqt-web==0.4.11
rqt_launch==0.4.9
rqt_nav_view==0.5.7
rqt_shell==0.4.11
rviz==1.14.25
scipy==1.15.3
sensor-msgs==1.13.1
simplejson==3.20.2
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smbus2==0.5.0
smclib==1.8.6
sniffio==1.3.1
srdfdom==0.6.4
sshkeyboard==2.3.1
starlette==0.49.3
sympy==1.14.0
tensorboardX==2.6.4
tensorstore==0.1.78
tf==1.13.3
tf-conversions==1.13.3
tf2-geometry-msgs==0.7.8
tf2-kdl==0.7.8
tf2-py==0.7.8
tf2-ros==0.7.8
toolz==1.1.0
topic-tools==1.17.0
torch==2.9.0+cu126
torchvision==0.24.0+cu126
tqdm==4.67.1
trac_ik_python==1.6.6
transforms3d==0.4.2
treescope==0.1.10
trimesh==4.10.1
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urdfdom-py==0.4.6
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
-e git+ssh://git@github.com/viberobotics/viberobotics-python.git@0ddc120cb67e1c30a053a006226eff3ab2ee2792#egg=viberobotics
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
xacro==1.14.20
zipp==3.23.0
